+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.31.0] - 2024-07-10
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): More properties of meal, e.g. is vegan, healthy and so on, and how much time it takes to prepare.
+
 ### Fixed
 - [Marek Kurańda](https://github.com/mjkuranda): Ingredient and meal images URL.